inventory-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
PyJWT 
psycopg2-binary==2.9.9
